Recalculate .keyboard-slider height when changing imgs to svgs. #15.
If UI is loaded on a smaller screen, the svg will be cut in half as the
height of the image will be smaller then the height of the svg. There
could be other solutions for this, but this was the quickest, and it's
sufficient for the purpose of the mockup.