fix: tooltip positioning (#2279)
This PR replaces all old circle icon with tooltip with the `circle-tooltip` component. This component tries to show the tooltip at bottom of the icon but if there is no enough place then render at the top of the icon.
The `circle-tooltip` also supports to set custom tooltip width so we don't need custom css classes.